Fix import skined mesh
By specification, skinned mesh object does not requires local transforms. In many cases it contains this transform, but there are files without it. In this case we can extract actual transform from binding matrices. We assume, that all joints has transforms that corresppond to transforms at the skinning momen. So, we gen the random joint, extract binding matrix and calculate global transform of the skinned mesh object.

If there are no binding matrices, then nothing to do, use (if it exists) transforms from node properties.