Export skinned object
We write to the deformer node properties transforms form StaticKineState. This parameter contains global transform of the null, which it hav at the envelope moment.
Transforms in the hierarchy corresponds to the default pose, but unskinned objects has actual transforms. So, after importing all skinned object will be at the origin and shifted with respect to it non-skinned parents.

If deformers are children of the scinned object, then it will import at origin. Because at the export these deformer's transforms calculated with respect to the parent (skinned mesh), and the transform of the skinned mesh in the import process calculated from the deformers.